Fix #14

+++ src/command.hpp
@@ -17,7 +17,7 @@ struct SimpleCommand {
 	void insertArgument(char * argument);
 	ssize_t numOfArguments = 0;
 	void release() {
-		for (size_t x = 0; x < arguments.size(); ++x) delete[] arguments[x];
+		for (int x = 0; x < arguments.size(); ++x) delete[] arguments[x];
 		arguments.clear();
 		arguments.shrink_to_fit();
 		numOfArguments = 0;
